Backport: Trunk 8305; quash a warning

+++ code/fred2/wing_editor.cpp
@@ -405,7 +405,7 @@ void wing_editor::initialize_data_safe(int full_update)
 		for (i=0; i<Wings[cur_wing].wave_count; i++)
 			ptr->AddString(Ships[Wings[cur_wing].ship_index[i]].ship_name);
 
-		m_threshold_spin.SetRange(0, Wings[cur_wing].wave_count - 1);
+		m_threshold_spin.SetRange(0, (short) Wings[cur_wing].wave_count - 1);
 		for (i=0; i<Num_reinforcements; i++)
 			if (!stricmp(Reinforcements[i].name, Wings[cur_wing].name))
 				break;
